requirements.txt
selenium>=4.10.0
webdriver-manager>=4.0.0
pytest>=7.0.0
pytest-xdist
pytest-html
allure-pytest
